update firebase_auth_facade with rustic + null safety

+++ lib/infrastructure/auth/firebase_auth_facade.dart
@@ -1,8 +1,10 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:dartz/dartz.dart';
 import 'package:firebase_auth/firebase_auth.dart' as auth;
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:injectable/injectable.dart';
+import 'package:rustic/option.dart';
+import 'package:rustic/result.dart';
+import 'package:rustic/tuple.dart';
 
 import 'package:wine/domain/auth/auth_failure.dart';
 import 'package:wine/domain/auth/email_address.dart';
@@ -24,7 +26,7 @@ class FirebaseAuthFacade implements IAuthFacade {
   final GoogleSignIn _googleSignIn;
 
   @override
-  Future<Either<AuthFailure, Unit>> convertWithEmailAndPassword(
+  Future<Result<Unit, AuthFailure>> convertWithEmailAndPassword(
     EmailAddress emailAddress,
     Password password,
   ) async {
@@ -44,20 +46,20 @@ class FirebaseAuthFacade implements IAuthFacade {
       currentUser = _firebaseAuth.currentUser;
       await currentUser?.sendEmailVerification();
 
-      return right(unit);
+      return const Ok(Unit());
     } on FirebaseException catch (e) {
       if (e.code == 'email-already-in-use') {
-        return left(const AuthFailure.emailAlreadyInUse());
+        return const Err(AuthFailure.emailAlreadyInUse());
       }
-      return left(const AuthFailure.serverError());
+      return const Err(AuthFailure.serverError());
     } catch (_) {
-      return left(const AuthFailure.unexpected());
+      return const Err(AuthFailure.unexpected());
     }
   }
 
   @override
-  Future<Option<User>> getLoggedInUser() async =>
-      optionOf(_firebaseAuth.currentUser?.toDomain());
+  Future<Option<User>?> getLoggedInUser() async =>
+      _firebaseAuth.currentUser?.toDomain().asOption();
 
   @override
   bool get isAnonymous {
@@ -72,19 +74,19 @@ class FirebaseAuthFacade implements IAuthFacade {
   }
 
   @override
-  Future<Either<AuthFailure, Unit>> logInAnonymously() async {
+  Future<Result<Unit, AuthFailure>> logInAnonymously() async {
     try {
       await _firebaseAuth.signInAnonymously();
-      return right(unit);
+      return const Ok(Unit());
     } on FirebaseException catch (_) {
-      return left(const AuthFailure.serverError());
+      return const Err(AuthFailure.serverError());
     } catch (_) {
-      return left(const AuthFailure.unexpected());
+      return const Err(AuthFailure.unexpected());
     }
   }
 
   @override
-  Future<Either<AuthFailure, Unit>> logInWithEmailAndPassword(
+  Future<Result<Unit, AuthFailure>> logInWithEmailAndPassword(
     EmailAddress emailAddress,
     Password password,
   ) async {
@@ -100,29 +102,27 @@ class FirebaseAuthFacade implements IAuthFacade {
         password: passwordStr,
       );
 
-      return right(unit);
+      return const Ok(Unit());
     } on FirebaseException catch (e) {
       await _firebaseAuth.signInAnonymously();
 
       if (e.code == 'wrong-password' || e.code == 'user-not-found') {
-        return left(
-          const AuthFailure.invalidEmailAndPasswordCombination(),
-        );
+        return const Err(AuthFailure.invalidEmailAndPasswordCombination());
       }
-      return left(const AuthFailure.serverError());
+      return const Err(AuthFailure.serverError());
     } catch (_) {
-      return left(const AuthFailure.unexpected());
+      return const Err(AuthFailure.unexpected());
     }
   }
 
   @override
-  Future<Either<AuthFailure, Unit>> logInWithGoogle() async {
+  Future<Result<Unit, AuthFailure>> logInWithGoogle() async {
     try {
       final anonymousUser = _firebaseAuth.currentUser;
 
       final googleAccount = await _googleSignIn.signIn();
       if (googleAccount == null) {
-        return left(const AuthFailure.cancelledByUser());
+        return const Err(AuthFailure.cancelledByUser());
       }
 
       final googleAuthentication = await googleAccount.authentication;
@@ -133,27 +133,27 @@ class FirebaseAuthFacade implements IAuthFacade {
       );
 
       try {
-        return right(await _trySignInWithGoogle(anonymousUser, authCredential));
+        return await _trySignInWithGoogle(anonymousUser, authCredential);
       } on FirebaseException catch (e) {
         if (e.code == 'credential-already-in-use') {
-          return right(await _signInWithGoogleCredentialAlreadyInUse(
+          return await _signInWithGoogleCredentialAlreadyInUse(
             anonymousUser,
             authCredential,
-          ));
+          );
         }
-        return left(const AuthFailure.serverError());
+        return const Err(AuthFailure.serverError());
       } catch (_) {
-        return left(const AuthFailure.unexpected());
+        return const Err(AuthFailure.unexpected());
       }
     } on FirebaseException catch (_) {
-      return left(const AuthFailure.serverError());
+      return const Err(AuthFailure.serverError());
     } catch (_) {
-      return left(const AuthFailure.unexpected());
+      return const Err(AuthFailure.unexpected());
     }
   }
 
   @override
-  Future<Either<AuthFailure, Unit>> logOut() async {
+  Future<Result<Unit, AuthFailure>> logOut() async {
     try {
       await Future.wait([
         _firebaseAuth.signOut(),
@@ -162,68 +162,76 @@ class FirebaseAuthFacade implements IAuthFacade {
 
       final currentUser = _firebaseAuth.currentUser;
       if (currentUser != null) {
-        return left(const AuthFailure.unableToSignOut());
+        return const Err(AuthFailure.unableToSignOut());
       }
 
       await _firebaseAuth.signInAnonymously();
-      return right(unit);
+      return const Ok(Unit());
     } on FirebaseException catch (_) {
-      return left(const AuthFailure.serverError());
+      return const Err(AuthFailure.serverError());
     } catch (_) {
-      return left(const AuthFailure.unexpected());
+      return const Err(AuthFailure.unexpected());
     }
   }
 
   @override
-  Future<Either<AuthFailure, Unit>> resendVerificationEmail() async {
+  Future<Result<Unit, AuthFailure>> resendVerificationEmail() async {
     try {
       final currentUser = _firebaseAuth.currentUser;
 
       if (currentUser != null) {
         await currentUser.sendEmailVerification();
-        return right(unit);
+        return const Ok(Unit());
       }
-      return left(const AuthFailure.unexpected());
+      return const Err(AuthFailure.unexpected());
     } on FirebaseException catch (_) {
-      return left(const AuthFailure.serverError());
+      return const Err(AuthFailure.serverError());
     } catch (_) {
-      return left(const AuthFailure.unexpected());
+      return const Err(AuthFailure.unexpected());
     }
   }
 
-  Future<Unit> _trySignInWithGoogle(
-    auth.User anonymousUser,
+  Future<Result<Unit, AuthFailure>> _trySignInWithGoogle(
+    auth.User? anonymousUser,
     auth.AuthCredential authCredential,
   ) async {
-    await anonymousUser.linkWithCredential(authCredential);
-    await _updateUserInfo(_googleSignIn.currentUser, anonymousUser);
+    if (anonymousUser != null) {
+      await anonymousUser.linkWithCredential(authCredential);
 
-    return unit;
+      return await _updateUserInfo(_googleSignIn.currentUser, anonymousUser);
+    }
+    return const Err(AuthFailure.unexpected());
   }
 
-  Future<Unit> _signInWithGoogleCredentialAlreadyInUse(
-    auth.User anonymousUser,
+  Future<Result<Unit, AuthFailure>> _signInWithGoogleCredentialAlreadyInUse(
+    auth.User? anonymousUser,
     auth.AuthCredential authCredential,
   ) async {
-    await anonymousUser.delete();
+    if (anonymousUser != null) {
+      await anonymousUser.delete();
 
-    await _firebaseAuth.signInWithCredential(authCredential);
+      await _firebaseAuth.signInWithCredential(authCredential);
 
-    var currentUser = _firebaseAuth.currentUser;
+      var currentUser = _firebaseAuth.currentUser;
 
-    await _updateUserInfo(_googleSignIn.currentUser, currentUser);
-
-    return unit;
+      return await _updateUserInfo(_googleSignIn.currentUser, currentUser);
+    }
+    return const Err(AuthFailure.unexpected());
   }
 
-  Future<void> _updateUserInfo(
-    GoogleSignInAccount currentAccount,
-    auth.User currentUser,
+  Future<Result<Unit, AuthFailure>> _updateUserInfo(
+    GoogleSignInAccount? currentAccount,
+    auth.User? currentUser,
   ) async {
-    await currentUser.updateProfile(
-      displayName: currentAccount.displayName,
-      photoURL: currentAccount.photoUrl,
-    );
-    await currentUser.reload();
+    if (currentAccount != null || currentUser != null) {
+      await currentUser?.updateProfile(
+        displayName: currentAccount?.displayName,
+        photoURL: currentAccount?.photoUrl,
+      );
+      await currentUser?.reload();
+
+      return const Ok(Unit());
+    }
+    return const Err(AuthFailure.unexpected());
   }
 }
